# 03-mvc/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "User API Documentation",
    "description": "",
    "version": "1.0.0"
  },
  "host": "localhost:8081/api/v0",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
    "/user": {
      "get": {
        "summary": "un résumé",
        "description": "une description",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Some description...",
            "schema": {
              "type": "object",
              "properties": {
                "login": {
                  "type": "string",
                  "example": "John"
                },
                "password": {
                  "type": "string",
                  "example": "pass"
                }
              },
              "required": [
                "login",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/user/{username}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/user/{username}/posts": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}